### 2021-09-10 #3715 @whedon
[PRE REVIEW]: PyLithics: A Python package for stone tool analysis
**Submitting author:** @JasonGellis (<a href="http://orcid.org/0000-0002-9929-789X">Jason Gellis</a>)
**Repository:** <a href="https://github.com/alan-turing-institute/Palaeoanalytics" target ="_blank">https://github.com/alan-turing-institute/Palaeoanalytics</a>
**Version:** v1.0
