### 2021-09-13 #3715 @JasonGellis
@arfon Hello, we have an open-source LICENSE.txt file in the repo but whedon can not detect it. There are links on the main README.md and in the paper file. Is there something wrong on our end?
