### 2021-09-13 #3715 @arfon
The license should be at the top level of the repository. It's currently nested in the `/paper` folder.
