### 2021-09-18 #3715 @Nikoleta-v3
👋 @MichaelHoltonPrice & @dombennett & @platipodium & @steko would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "PyLithics: A Python package for stone tool analysis". You can find more information at the top of this Github issue 🆙
