### 2021-09-21 alan-turing-institute/Palaeoanalytics#129 @steko
License
Hello! I started the JOSS review as outlined here https://github.com/openjournals/joss-reviews/issues/3738 and I think the license needs some clarifying/cleanup.

The repository contains references to various different licenses:

- there is a `license` file at the root of the repository, Creative Commons Attribution-ShareAlike 4.0
- the README has a badge linking to https://creativecommons.org/licenses/by-nc-sa/4.0/ Creative Commons Attribution-**NonCommercial**-ShareAlike 4.0
- `setup.py` has "MIT" in the license descriptor
- `paper/LICENSE.txt` is again verbatim Creative Commons Attribution-ShareAlike 4.0

I think it would be better to clarify which license applies to each specific part of the repository or just adopt one license for all.

I should note that [Creative Commons advises against using their own licenses for software](https://creativecommons.org/faq/#can-i-apply-a-creative-commons-license-to-software) and that CC licenses are not [approved the the Open Source Initiative](https://opensource.org/licenses/alphabetical).

I don't know if you intended to differentiate the license of the software from that of the paper, but it doesn't look like JOSS guidelines encourage applying any specific license to the paper.

I hope I described the issue in sufficient detail.
