### 2021-09-23 cec2b8c
Add complexity measurement function
