### 2021-09-23 430ef40
Update license information to GNU GPLv3
