### 2021-09-23 e351373
Update license information to GNU GPLv3
