### 2021-09-23 819d12e
Update license information to GNU GPLv3
