+++ LICENSE.txt
@@ -632,7 +632,7 @@ state the exclusion of warranty; and each file should have at least
 the "copyright" line and a pointer to where the full notice is found.
 
     PyLithics: A Python package for stone tool analysis
-    Copyright (C) 2021 Jason Jacob Gellis and Camila Rangel Smith
+    Copyright (C) 2021 Jason Jacob Gellis, Camila Rangel Smith. and Robert A. Foley
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -652,7 +652,7 @@ Also add information on how to contact you by electronic and paper mail.
   If the program does terminal interaction, make it output a short
 notice like this when it starts in an interactive mode:
 
-    PyLithics: A Python package for stone tool analysis  Copyright (C) 2021  Jason Jacob Gellis and Camila Rangel Smith
+    PyLithics: A Python package for stone tool analysis  Copyright (C) 2021 Jason Jacob Gellis, Camila Rangel Smith. and Robert A. Foley
     This program comes with ABSOLUTELY NO WARRANTY; for details type `show w'.
     This is free software, and you are welcome to redistribute it
     under certain conditions; type `show c' for details.
