### 2021-09-24 alan-turing-institute/Palaeoanalytics#129 @JasonGellis
@steko Hi Stefano, I've updated the licens across the repo and merged to the develop branch. If, after review, this meets your approval I will merge to the main branch.
