### 2021-09-29 d9090cb
adding test for complexity measure
