### 2021-09-30 7b8d12d
adding function that plots complexity
