+++ pylithics/src/plotting.py
@@ -191,6 +191,10 @@ def plot_results(id, image_array, contours_df, output_dir):
     output_lithic = os.path.join(output_dir, id + "_lithium_angles.png")
     plot_angles(image_array, contours_df, output_lithic)
 
+    # plot scar strike angle
+    output_lithic = os.path.join(output_dir, id + "_complexity_polygon_count.png")
+    plot_complexity(image_array, contours_df, output_lithic)
+
 
 
 def plot_thresholding(image_array, threshold, binary_array, output_file=''):
@@ -279,3 +283,43 @@ def plot_template_arrow(image_array, template_array, value):
     ax[1].set_yticks([])
     plt.figtext(0.4, 0.9, str(value))
     plt.show()
+
+def plot_complexity(image_array, contours_df, output_path):
+    """
+    Plot the contours from the lithic surfaces and display complexity and polygon count measurements.
+
+    Parameters
+    ----------
+    image_array: array
+        Original image array (0 to 255)
+    contours_df: dataframe
+        Dataframe with detected contours and extra information about them.
+    output_path: str
+        Path to output directory to save processed images
+
+    """
+    fig_x_size = fig_size(image_array)
+    fig, ax = plt.subplots(figsize=(fig_x_size, 20))
+    ax.imshow(image_array, cmap=plt.cm.gray)
+
+    # selecting only scars with a complexity measure > 0
+    contours_angles_df = contours_df[(contours_df['parent_index'] != -1) & (contours_df['complexity']>0)]
+    cmap_list = plt.cm.get_cmap('tab20', contours_angles_df.shape[0])
+
+    if contours_angles_df.shape[0] == 0:
+        warnings.warn("Warning: No scars with complexity measure, no complexity output figure will be saved.'")
+        return None
+
+    i = 0
+    for contour, complexity, polygon_count in \
+            contours_angles_df[['contour', 'complexity','polygon_count']].itertuples(index=False):
+        text = "Complexity: " + str(complexity)+", Polygon Count: "+str(polygon_count)
+        ax.plot(contour[:, 0], contour[:, 1], label=text, linewidth=5, color=cmap_list(i))
+        i = i + 1
+
+    ax.set_xticks([])
+    ax.set_yticks([])
+    plt.legend(bbox_to_anchor=(1.02, 0), loc="lower left", borderaxespad=0, fontsize=11)
+    plt.title("Scar complexity and polygon count measurements", fontsize=30)
+    plt.savefig(output_path)
+    plt.close(fig)
