### 2021-09-30 d88dc1a
changing minimum complexity threshold
