### 2021-09-30 3770295
fixing broken test and removing unsued column in contour dataframe
