+++ tests/test_pipeline.py
@@ -38,7 +38,7 @@ def test_pipeline():
     json_output = data_output(contours, config_file)
 
     assert len(json_output) == 4
-    assert contours.shape == (11, 17)
+    assert contours.shape == (11, 16)
     assert binary_array.shape == (1841, 1665)
     assert len(json_output['lithic_contours']) == 4
 
@@ -76,5 +76,5 @@ def test_arrow_pipeline():
     contours_final = get_scars_angles(image_processed, contours, arrow_df)
 
     assert len(templates) == 4
-    assert contours_final.shape == (11, 15)
+    assert contours_final.shape == (11, 16)
     assert arrow_df.shape == (4, 2)
