### 2021-09-30 alan-turing-institute/Palaeoanalytics#129 @JasonGellis
@steko Hi Stefano, we addressed this issue in PR #133 and have now merged into the main branch. We will now close this issue.
