### 2021-10-03 #3738 @Nikoleta-v3
Yup @steko 👍🏻

Working your way through the checklist and opening issues if you think that the paper hasn't filled the requirements is exactly what we want.

If you feel that some points require more discussion  feel free to share your thoughts and comments here, and also feel free to share your overall thoughts about the paper.

I hope I have answered your question 😄
