### 2021-11-10 #3738 @MichaelHoltonPrice
@steko and I have raised some issues that, speaking for myself at least, need to be addressed before finishing the review. None are major, but they do prevent me from marking some items in the checklist.
