### 2021-11-11 47b591e
Joss corrections issues #136, 137, & 138
