+++ paper/paper.bib
@@ -1,22 +1,29 @@
 
-@InProceedings{ mckinney-proc-scipy-2010,
-  author    = { {W}es {M}c{K}inney },
-  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
-  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
-  pages     = { 56 - 61 },
-  year      = { 2010 },
-  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
-  doi       = { 10.25080/Majora-92bf1922-00a }
+@InProceedings{McKinney-proc-scipy:2010,
+  author    = {{W}es {M}c{K}inney},
+  title     = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
+  booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
+  pages     = {56 - 61},
+  year      = {2010},
+  editor    = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
+  doi       = { 10.25080/Majora-92bf1922-00a}
 }
 
-@article{2020NumPy-Array,
-  author  = {Harris, Charles R. and Millman, K. Jarrod and van der Walt, Stéfan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and van Kerkwijk, Marten H. and Brett, Matthew and Haldane, Allan and Fernández del Río, Jaime and Wiebe, Mark and Peterson, Pearu and Gérard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
-  title   = {Array programming with {NumPy}},
-  journal = {Nature},
-  year    = {2020},
-  volume  = {585},
-  pages   = {357–362},
-  doi     = {10.1038/s41586-020-2649-2}
+@article{Harris et al.:2020,
+      author = {Charles R. Harris and K. Jarrod Millman and Stéfan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fernández del Río and Mark Wiebe and Pearu Peterson and Pierre Gérard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
+   doi = {10.1038/s41586-020-2649-2},
+   issn = {1476-4687},
+   issue = {7825},
+   journal = {Nature 2020 585:7825},
+   keywords = {Computational neuroscience,Computational science,Computer science,Software,Solar physics},
+   month = {9},
+   pages = {357-362},
+   pmid = {32939066},
+   publisher = {Nature Publishing Group},
+   title = {Array programming with NumPy},
+   volume = {585},
+   url = {https://www.nature.com/articles/s41586-020-2649-2},
+   year = {2020},
 }
 @article{opencv_library,
     author = {Bradski, G.},
@@ -28,7 +35,7 @@
     title = {{The OpenCV Library}},
     year = {2000}
 }
-@article{2020SciPy-NMeth,
+@article{Virtanen et al.:2020,
[…]
