### 2021-11-11 8eb9952
Joss corrections issues #136, 137, & 138
