+++ README.md
@@ -1,6 +1,6 @@
 <h1 align="center">Welcome to Palaeoanalytics! </h1>
 
-> Repository for the [Paleoanalytics project](https://www.turing.ac.uk/research/research-projects/palaeoanalytics). 
+> Repository for the [Palaeoanalytics project](https://www.turing.ac.uk/research/research-projects/palaeoanalytics). 
 >A collaboration between The Alan Turing Institute and the University of Cambridge.  
 
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
@@ -54,7 +54,7 @@ The inputs for `PyLithics` are images of lithic objects, images of their associa
 
 `PyLithics` processes the images with the following steps (and as illustrated in the schema below):
 
-1. Import images and match image name find to associated image ID and scale image from CSV meetadata file.
+1. Import images and match to associated image ID and scale image from CSV metadata file.
 2. Calculate a conversion of pixels to millimeters based on the size of the associated scale from CSV metadata file. If no scale is present, measurements will be in pixels
 3. Apply noise removal and contrast stretching to images to minimise pixel variation.
 4. Pixel intensity thresholding of images to prepare for contour finding.
@@ -74,29 +74,28 @@ Here you can find a schema of the workflow described above:
 ## Installation
 The `PyLithics` package requires Python 3.7 or greater. To install, start by creating a fresh conda environment.
 ```
-conda create -n paleo python=3.7
-conda activate paleo
+python3 -m venv palaeo
+source palaeo/bin/activate
 ```
 
-Get the source.
+Clone the repository.
 ```
 git clone https://github.com/alan-turing-institute/Palaeoanalytics.git
 ```
 
-Enter the repository and check out a relevant branch if necessary (the `develop` branch contains the most up to date stable version of the code, but this branch is fast moving
-if you want to have a stable and static version is better to use `main` branch).
+Enter the repository and check out a relevant branch if necessary (the `develop` branch contains the most up-to-date stable version of the code, but this branch is fast moving.
+If you want to have a stable and static version it is better to use `main` branch).
 ```
 cd Palaeoanalytics
 git checkout main
 ```
-Install OpenCV using conda and the rest of packages using `pip`.
+Install 'PyLithics'.
 ```
 pip install .
 ```
 
-**Note**: For Mac users we recommend to use an OS versions=> 10.14 to prevent build problems. For some OS (e.g. Mojave OSX), the OpenCV installation can take too long, if this is the case you can also install
-the library from conda using the line `conda install -c conda-forge opencv=4.5.3`. Also, be careful about versions,
-older OpenCV versions can break this code. 
+**Note**: For Mac users we recommend an OS versions=> 10.14 to prevent build problems. For some OS (e.g. Mojave OSX), the OpenCV installation can take too long, if this is the case you can also install
+the library from conda using the line `pip install opencv=4.5.4.58`. Also, be careful about versions, as older versions of OpenCV can break this code. 
 
 ## Running `PyLithics`
 
@@ -180,7 +179,7 @@ This CSV file should have as a minimum the following 3 variables:
 - *scale_ID*: The scale id (name of the scale image file)
 - *PA_scale*: The scale measurement (how many centimeters this scale represents).
 
-An example of this table, where one scale correspond to several images is the following:
+An example of this table, where one scale corresponds to several images is the following:
 
 |PA_ID | scale_ID  | PA_scale  | 
 |------|-----------|-----------|
@@ -197,7 +196,7 @@ with the command:
 ```
 pylithics_run -c configs/test_config.yml --input_dir <path_to_input_dir> --output_dir <path_to_output_directory> 
 ```
-lithics image files must still be inside  the '<path_to_input_dir>/images/' directory). However, all the measurements will only be
[…]
