### 2021-11-12 5e8a221
Minor corrections for issues #138
