### 2021-11-12 1d444ae
Minor corrections for issues #138
