### 2021-11-12 eeec0e3
Minor corrections for issues #138
