+++ requirements.txt
@@ -4,5 +4,5 @@ pytest
 pandas
 scikit-image
 scipy
-opencv-python==4.5.4.58
+opencv-python
 pyyaml==5.4.1
