### 2021-11-16 f9b8e4b
adding opencv-contrib-python library
