+++ .travis.yml
@@ -1,22 +1,14 @@
 matrix:
  include:
-  - language: python
-    python:
-     - "3.7"
-    install:
-     - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
-     - bash miniconda.sh -b -p $HOME/miniconda
-     - export PATH="$HOME/miniconda/bin:$PATH"
-     - export TRAVIS=true
-     - hash -r
-     - conda config --set always_yes yes --set changeps1 no
-     - conda update -q conda
-     - conda info -a
-     - conda create -q -n test-env python=$TRAVIS_PYTHON_VERSION
-     - source activate test-env
-     - pip install .
-    script:
-     - pytest -s
+  language: python
+  python:
+   - "3.7"
+  # command to install dependencies
+  install:
+   - pip install .
+  # command to run tests
+  script:
+   - pytest -s
 
 after_failure:
  - cat config.log
