+++ .travis.yml
@@ -5,6 +5,7 @@ matrix:
    - "3.7"
   # command to install dependencies
   install:
+   - pip install --upgrade pip
    - pip install .
   # command to run tests
   script:
