### 2021-11-18 713a320
Minor corrections for issues #138
