### 2021-11-18 20b32ed
Minor corrections for issues #139
