+++ README.md
@@ -94,10 +94,10 @@ Install 'PyLithics'.
 pip install .
 ```
 
+
 **Note**: For Mac users we recommend an OS versions=> 10.14 to prevent build problems.
 
 ## Running `PyLithics`
-
 
 `PyLithics` can be run via command line. The following command displays all available options:
 
