### 2021-11-18 f9c2ad8
Minor corrections for README
