### 2021-11-19 f3771db
Minor corrections for README
