### 2021-11-19 19ff37a
Minor corrections for README
