### 2021-11-19 88cfdeb
Minor corrections for README
