### 2021-11-19 9f68ddf
Minor corrections for README
