### 2021-11-19 b8c5b11
Minor corrections for JOSS
