+++ setup.py
@@ -4,16 +4,19 @@ with open("requirements.txt", "r") as f:
     REQUIRED_PACKAGES = f.read().splitlines()
 
 setup(
-    name="pylithics",
+    name="PyLithics",
     version="1.0",
     description="A Python package for stone tool analysis",
-    url="https://github.com/alan-turing-institute/Palaeoanalytics",
+    download_url="https://github.com/alan-turing-institute/Palaeoanalytics",
     author="Jason Gellis, Camila Rangel Smith, Robert Foley",
+    author_email='jg760@cam.ac.uk',
     license="GNU GPLv3",
     include_package_data=True,
     packages=["pylithics",
               "pylithics.src",
               "pylithics.scripts"],
+    keywords = ['lithics', 'human evolution', 'Archaeology', 'archeology',
+                'lithic analysis', 'prehistoric technology', 'computer vision'],
     install_requires=REQUIRED_PACKAGES,
     # we will need this later, I'll leave it commented as cont reminder.
     entry_points={"console_scripts": [
