### 2021-11-19 30a2c52
Minor corrections for JOSS
