### 2021-11-19 fccb0ed
Minor corrections for JOSS
