### 2021-11-19 569b8ad
Minor corrections for JOSS
