### 2021-11-19 8bb13fc
Minor corrections for JOSS
