### 2021-11-19 cec0e35
Minor corrections for JOSS
