### 2021-11-19 7cf0839
Minor corrections for JOSS
