### 2021-11-19 9761d66
Minor corrections for JOSS
