### 2021-11-19 6475e0c
Minor corrections for JOSS
