### 2021-11-19 e20244d
Minor corrections for JOSS
