### 2021-11-19 a0d08c7
Minor corrections for JOSS
