### 2021-11-19 d7da3e4
Minor corrections for JOSS
