### 2021-11-19 dc31659
Minor corrections for JOSS
