### 2021-11-19 c85a8cf
Minor corrections for JOSS
