### 2021-11-19 da3e6b5
Minor corrections for JOSS
