+++ README.md
@@ -1,51 +1,107 @@
-# Palaeoanalytics
-Repository for the Paleoanalytics project.
+<h1 align="center">Welcome to Palaeoanalytics! </h1>
 
-For over one hundred years, archaeologists have recovered, recorded, photographed and drawn stone tools, and used these to build up strong models of their
-typologies, technologies, distributions and affinities. Much of this information is recorded in publications of greater or lesser accessibility, and large-scale,
-comprehensive databases are few and far between. To improve the scale and quality of these data, the PALAEONALYTICS Project (https://www.turing.ac.uk/research/research-projects/palaeoanalytics) has utilized computer vision methodologies to develop PyLithics, an open-source, free for use, software package to capture rapidly large volumes of these data in usable forms. PyLithics v1.0, processes lithic artefact illustrations scanned from the literature, and has been optimised for feature extraction and measurement using a number of computer vision techniques including pixel intensity thresholding, edge detection and segmentation, and custom template matching and image kernels. 
- 
-On both conventional and modern drawings, PyLithics v1.0 can identify and measure dorsal surface scar number, shape, size, orientation and diversity. The
-orientation and flaking direction of dorsal scars can also be calculated from either ripples or arrows. Complete size and shape metrics of individual scars and
-whole flakes can be calculated and recorded in a customisable database. The resulting data can be used for metrical analysis, extracting features indicative of
-both typologies and technological processes. Data output has been optimised for machine learning and neural network processing algorithms and can easily be
-employed to explore patterns of variation within and between assemblages.  
+> Repository for the [Palaeoanalytics project](https://www.turing.ac.uk/research/research-projects/palaeoanalytics). 
+> A collaboration between The Alan Turing Institute and the University of Cambridge.  
 
-# Installation
+[![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
+[![Build Status](https://app.travis-ci.com/alan-turing-institute/Palaeoanalytics.svg?token=sMJzQpXKRs31ujsqXNxP&branch=develop)](https://app.travis-ci.com/alan-turing-institute/Palaeoanalytics)
 
-The `pylithics` package requires Python 3.6 or greater. To install, start by creating a fresh conda environment.
+# **Table of Contents:**
+
+- [About the project](#about)
+- [The team](#team)
+- [The `PyLithics` package](#pylithics)
+- [Drawing style for PyLithics](#drawing)
+- [Contributing](#contributing)
+- [Licence](#licence)
+
+# 📖 About the project <a name="about"></a>
+
+Archaeologists have long used stone tools (lithics) to reconstruct the behavior of prehistoric hominins. While techniques 
+have become more quantitative, there still remain barriers to optimizing data retrieval. Machine learning and computer 
+vision approaches can be developed to extract quantitative and trait data from lithics, photographs and drawings. `PyLithics`
+has been developed to capture data from 2D line drawings, focusing on the size, shape and technological attributes of flakes. 
+
+`PyLithics`is an open-source, free for use software package for processing lithic artefact illustrations scanned from 
+the literature. This tool accurately identifies, outlines, and computes lithic shape and linear measures, and returns user 
+ready data. It has been optimized for feature extraction and measurement using a number of computer vision techniques 
+including pixel intensity thresholding, edge detection, contour finding, custom template matching and image kernels. 
+On both conventional and modern drawings, `PyLithics`can identify and platform, lateral, dorsal, and ventral surfaces,
[…]
