### 2021-11-19 bad612d
Minor corrections for JOSS
