### 2021-11-24 #3738 @Nikoleta-v3
👋🏻 @JasonGellis `whedon` failed to compile the pdf because of an error in the `bib` file. Can you please tweak the file?
