### 2021-11-26 dcdf39d
Minor corrections for JOSS
