### 2021-11-26 d9e8abb
Minor corrections for JOSS
