### 2021-11-26 62dd70f
Minor corrections for JOSS
