+++ README.md
@@ -1,7 +1,7 @@
 <h1 align="center">Welcome to Palaeoanalytics! </h1>
 
 > Repository for the [Palaeoanalytics project](https://www.turing.ac.uk/research/research-projects/palaeoanalytics). 
->A collaboration between The Alan Turing Institute and the University of Cambridge.  
+> A collaboration between The Alan Turing Institute and the University of Cambridge.  
 
 [![License: GPL v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 [![Build Status](https://app.travis-ci.com/alan-turing-institute/Palaeoanalytics.svg?token=sMJzQpXKRs31ujsqXNxP&branch=develop)](https://app.travis-ci.com/alan-turing-institute/Palaeoanalytics)
@@ -16,24 +16,25 @@
 - [Licence](#licence)
 
 # 📖 About the project <a name="about"></a>
+
 Archaeologists have long used stone tools (lithics) to reconstruct the behavior of prehistoric hominins. While techniques 
 have become more quantitative, there still remain barriers to optimizing data retrieval. Machine learning and computer 
 vision approaches can be developed to extract quantitative and trait data from lithics, photographs and drawings. `PyLithics`
 has been developed to capture data from 2D line drawings, focusing on the size, shape and technological attributes of flakes. 
 
-`PyLithics`is an open-source, free for use, software package for processing lithic artefact illustrations scanned from 
-the literature. This tool accurately identifies, outlines and computes lithic shape and linear measures, and returns user 
+`PyLithics`is an open-source, free for use software package for processing lithic artefact illustrations scanned from 
+the literature. This tool accurately identifies, outlines, and computes lithic shape and linear measures, and returns user 
 ready data. It has been optimized for feature extraction and measurement using a number of computer vision techniques 
 including pixel intensity thresholding, edge detection, contour finding, custom template matching and image kernels. 
-On both conventional and modern drawings, `PyLithics`can identify and platform, lateral, dorsal, and ventral surfaces,
+On both conventional and modern drawings, `PyLithics` can identify and label platform, lateral, dorsal, and ventral surfaces,
 as well as individual dorsal surface scar shape, size, orientation, diversity, number, and flaking order. Complete size
 and shape metrics of individual scars and whole flakes can be calculated and recorded. Orientation and flaking direction 
 of dorsal scars can also be calculated. The resulting data can be used for metrical analysis, extracting features indicative
-of both typologies and technological processes. Data output can easily be employed to explore patterns of variation within and between assemblages.
+of typologies and technological processes. Data output can easily be employed to explore patterns of variation within and between assemblages.
 
 # 👥 The team <a name="team"></a>
 
-These are the members of the Palaoanalytics team as updated August 2021:
+These are the members of the Palaeoanalytics team as updated August 2021:
 
 | Name | Role | email | Github | 
 | --- | --- | --- | --- |
@@ -41,20 +42,19 @@ These are the members of the Palaoanalytics team as updated August 2021:
 | Camila Rangel Smith | Research Data Scientist (The Alan Turing Institute) | [crangelsmith@turing.ac.uk](mailto:crangelsmith@turing.ac.uk) |[@crangelsmith](https://github.com/crangelsmith) |
 | Robert Foley | Principal Investigator (University of Cambridge) | [raf10@cam.ac.uk](mailto:raf10@cam.ac.uk)| [Rob-LCHES](https://github.com/Rob-LCHES)
 
-
 # 📦 The `PyLithics` package <a name="pylithics"></a>
 
 > PyLithics: A Python package for stone tool analysis
 
 ## Workflow
 
[…]
