+++ README.md
@@ -243,4 +243,3 @@ All contributions will have to pass these automated tests to be merged into the 
 # 📝 Licence <a name="licence"></a>
 
 This software is licensed under the terms of the [GNU General Public License v3.0 (GNU GPLv3)](https://choosealicense.com/licenses/gpl-3.0/).
-
