+++ paper/paper.bib
@@ -1,108 +1,112 @@
 
-
-@article{McKinney-proc-scipy:2010,
-     ,author = {Wes McKinney},
-     title = {Data Structures for Statistical Computing in Python},
-     book_title = {Proceedings of the 9th Python in Science Conference},
-     pages = {56 - 61},
-     year = {2010},
-     editor = {St'efan van der Walt and Jarrod Millman},
-     doi = {10.25080/Majora-92bf1992-00a}
+@InProceedings{McKinney-proc-scipy:2010,
+    author = {{W}es {M}c{K}inney},
+    title = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
+    booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
+    pages = {56-61},
+    year = {2010},
+    editor = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
+    doi = { 10.25080/Majora-92bf1922-00a}
 }
 
-@article{Harris:2020,
+@article{Harris et al.:2020,
     author = {Charles R. Harris and K. Jarrod Millman and Stéfan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fernández del Río and Mark Wiebe and Pearu Peterson and Pierre Gérard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
-    doi = {10.1038/s41586-020-2649-2},
-    issn = {1476-4687},
-    issue = {7825},
-    journal = {Nature 2020 585:7825},
-    month = {9},
-    pages = {357-362},
-    pmid = {32939066},
-    publisher = {Nature Publishing Group},
-    title = {Array programming with NumPy},
-    volume = {585},
-    url = {https://www.nature.com/articles/s41586-020-2649-2},
-    year = {2020}
+   doi = {10.1038/s41586-020-2649-2},
+   issn = {1476-4687},
+   issue = {7825},
+   journal = {Nature 2020 585:7825},
+   keywords = {Computational neuroscience,Computational science,Computer science,Software,Solar physics},
+   month = {9},
+   pages = {357-362},
+   pmid = {32939066},
+   publisher = {Nature Publishing Group},
+   title = {Array programming with NumPy},
+   volume = {585},
+   url = {https://www.nature.com/articles/s41586-020-2649-2},
+   year = {2020},
 }
-
 @article{opencv_library,
     author = {Bradski, G.},
     citeulike-article-id = {2236121},
     journal = {Dr. Dobb's Journal of Software Tools},
+    keywords = {bibtex-import},
     posted-at = {2008-01-15 19:21:54},
     priority = {4},
     title = {{The OpenCV Library}},
     year = {2000}
 }
-
-@article{Virtanen:2020,
[…]
