### 2021-11-26 df19f4c
Updates to bib file
