+++ paper/paper.bib
@@ -1,112 +1,110 @@
 
-@InProceedings{McKinney-proc-scipy:2010,
-    author = {{W}es {M}c{K}inney},
-    title = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
-    booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
-    pages = {56-61},
-    year = {2010},
-    editor = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
-    doi = { 10.25080/Majora-92bf1922-00a}
+@inproceedings{McKinney:2010,
+    author = {McKinney, Wes},
+    booktitle = {Proceedings of the 9th Python in Science Conference},
+    doi = {10.25080/Majora-92bf1922-00a},
+    pages = {56--61},
+    title = {{Data Structures for Statistical Computing in Python}},
+    url = {https://conference.scipy.org/proceedings/scipy2010/mckinney.html},
+    year = {2010}
 }
 
-@article{Harris et al.:2020,
-    author = {Charles R. Harris and K. Jarrod Millman and Stéfan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fernández del Río and Mark Wiebe and Pearu Peterson and Pierre Gérard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
-   doi = {10.1038/s41586-020-2649-2},
-   issn = {1476-4687},
-   issue = {7825},
-   journal = {Nature 2020 585:7825},
-   keywords = {Computational neuroscience,Computational science,Computer science,Software,Solar physics},
-   month = {9},
-   pages = {357-362},
-   pmid = {32939066},
-   publisher = {Nature Publishing Group},
-   title = {Array programming with NumPy},
-   volume = {585},
-   url = {https://www.nature.com/articles/s41586-020-2649-2},
-   year = {2020},
+@article{Virtanen:2020,
+    author = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and van der Walt, St{\'{e}}fan J. and Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and Kern, Robert and Larson, Eric and Carey, C J and Polat, İlhan and Feng, Yu and Moore, Eric W. and VanderPlas, Jake and Laxalde, Denis and Perktold, Josef and Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and Harris, Charles R. and Archibald, Anne M. and Ribeiro, Ant{\^{o}}nio H. and Pedregosa, Fabian and van Mulbregt, Paul},
+    doi = {10.1038/s41592-019-0686-2},
+    issn = {1548-7091},
+    journal = {Nature Methods},
+    month = {mar},
+    number = {3},
+    pages = {261--272},
+    title = {{SciPy 1.0: fundamental algorithms for scientific computing in Python}},
+    url = {http://www.nature.com/articles/s41592-019-0686-2},
+    volume = {17},
+    year = {2020}
 }
+
+@article{Harris:2020,
+    archivePrefix = {arXiv},
+    arxivId = {2006.10256},
[…]
