### 2021-11-26 3ab130f
Updates to paper and bib file
