### 2021-11-26 a7616e3
Updates to paper and bib file
