### 2021-11-26 79cc5fc
Updates to JOSS paper
