### 2021-11-26 #3738 @JasonGellis
@Nikoleta-v3 The bib file has been fixed and the PDF can be generated. I have also addressed all issues raised by @steko and @MichaelHoltonPrice and merged them into the main branch of the repo.
