### 2021-12-05 #3738 @steko
Hi, and apologies for the delay. I still have two comments about the paper that aren't specific to any of the checks but I think would help improve the quality and clarity of the final version.

1. be more explicit not only about the input images that work well, but also about the drawing styles that don't work - in my experience a lot of lithics drawings have more details than just the outline of scars and your tool does a good job but not optimal, e.g.

![sanremo1](https://user-images.githubusercontent.com/40545/144757962-ddb032ad-3e22-471f-a303-4b34c605af8e.png)

![sanremo1_lithium_scars](https://user-images.githubusercontent.com/40545/144757996-e38ceb0b-89be-4b54-822f-f67237c0581a.png)

2. the JSON output is incredibly rich but not documented in the paper, you should at least inlcude what is now in `output_example.md` file already in the repository - and maybe give some example of how to process the JSON data for further analysis?
