### 2021-12-08 #3738 @JasonGellis
@steko @MichaelHoltonPrice @Nikoleta-v3 Thank you for your feedback. While differences in drawing styles are addressed in greater detail on the README.md of the repo, we agree that it belongs in the paper too, and this has been added.

We agree that that an example output of the JSON file should be included, and a new figure has been added (Figure 5). However, we feel that further discussion of how the data should be processed and analysed should be left to the user and does not belong in this paper. Metric output data and surface identification data are included in (current) figures 2 and 3. We feel that the addition of the JSON file to these two figures will adequately inform users/readers of the type of data PyLithics will produce.

All requests have been updated and merged to the main branch
