### 2021-12-19 #3738 @Nikoleta-v3
Hey @JasonGellis  thank you for your patience. I have reached out to the reviewer via email. I have seen that you have addressed their comments but I would like to wait a little bit more from them to confirm the changes.

Some minor comments from me:

- **Regarding the repository.** The contributing file is excellent, but it's missing some guidelines on running the test suite. Could you please add the information.

- **Regarding the paper.**
    - Line 61 typo `arrow` -> `arrows`
