### 2022-01-04 85e78b0
Updates to JOSS paper
