### 2022-01-04 #3738 @MichaelHoltonPrice
Apologies, the notifications are not always informative (e.g., I don't
think you realized I was waiting for updates from your team on such
things as the test script for some time). Nikoleta let me know that the
manuscript was ready for me again Dec 19th but I have been on vacation
since then (and still am). I can complete my review by the end of next week.

On Tue, Jan 4, 2022 at 8:43 AM Jason Gellis ***@***.***>
wrote:

> Happy New Year @Nikoleta-v3 <https://github.com/Nikoleta-v3> I have
> addressed the above issues and merged into main. Can I request a new editor
> as @MichaelHoltonPrice <https://github.com/MichaelHoltonPrice> has not
> been in contact since October.
>
