### 2022-01-13 alan-turing-institute/Palaeoanalytics#167 @MichaelHoltonPrice
Either note operating systems for which the README applies or add Windows specific commands
Windows requires a slightly different command to activate the environment:

.\palaeo\Scripts\activate

Also, it is sometimes necessary, especially if using Powershell, to first  set execution policy scopes:

Set-ExecutionPolicy Unrestricted -Scope Process
.\palaeo\Scripts\activate

I realize you may not want to support all operating systems (it can be a headache) and if so you might simply note so.
