### 2022-01-14 alan-turing-institute/Palaeoanalytics#169 @MichaelHoltonPrice
A means for running tests during dev with corresponding directions
Again, my commendations for a remarkable piece of work. One hurdle I faced was confirming that the tests in fact ran. I'm sure Travis continuous integration is a fine tool to use for testing, but it is, apparently, not easy to use during dev (or perhaps it is yet I couldn't figure out how to do so). The directions I found online suggested using Docker for local development. While I'm quite familiar with Docker, it seemed more hassle than it's worth. My guess is that during your development work you must have a way of running your tests before issuing a pull request. I suggest adding directions to the README, though I am not going to hold up publication for this.

In the end, the way I ran the tests was with the following script run from the root of the github repository. This is obviously incredibly "klugy". Despite being a seasoned programmer, however, I couldn't figure out a better way. Likely this is just ignorance on my part on how Travis CI works, but I did spend a few hours trying to figure out how to run the tests locally the "correct" way and all I was able to find were suggestions to use Docker, which I realized would be more work to do than just writing the following simple script.

import os

test_files = os.listdir('tests')

test_files = [f for f in test_files if f.endswith('.py')]

for f in test_files:
    print('----')
    print('Running tests in ' + f)
    file_stem = f.split('.')[0]
    exec('from tests import ' + file_stem)
    exec('functions = dir(' + file_stem + ')')
    functions = [f for f in functions if f.startswith('test_')]
    for func in functions:
        print('--')
        print('Calling ' + func)
        exec(file_stem + '.' + func + '()')
