### 2022-01-17 #3738 @Nikoleta-v3
Hey @JasonGellis 👋🏻  MichaelHoltonPrice opened three last issues.  They made it clear that they recommend the article for publication and that these issues are sufficient to hold up publication of the article.

Nevertheless, I would you like you to address issue [#167](https://github.com/alan-turing-institute/Palaeoanalytics/issues/167) and issue [#169](https://github.com/alan-turing-institute/Palaeoanalytics/issues/169). Issue 169 is similar to my comment which I can see you addressed by including information regarding Travis in the README file. Could you also include guidelines on how one can run the tests locally on their machines before pushing on GitHub?
