### 2022-01-19 4f03bfb
Add info for creating a venv in Windows OS.
