### 2022-01-19 alan-turing-institute/Palaeoanalytics#167 @JasonGellis
@MichaelHoltonPrice  The information for creating a virtual environment in Windows as been added to the README and merged to the main branch.
