### 2022-01-21 alan-turing-institute/Palaeoanalytics#169 @JasonGellis
@MichaelHoltonPrice @Nikoleta-v3
Test coverage documentation has been completed. I think there might me some confusion on this. We use the the pytest library. All that's needed is to go to the main Palaeoanalytics directory and type `pytest -s`

I have added a testing section section - Functional testing of PyLithics - in the REAME.md
