### 2022-01-23 #3738 @Nikoleta-v3
👋🏻  @JasonGellis. I `missing DOI` that `whedon` is recommending is the DOI to a review of the book. So please ignore the message (feel free to also check).

Anyway, everything looks good to me!

At this point could you:

- [ ] Make a  new tagged release of your software if you think that it is needed (current one is v1.0), and list the version tag of the archived version here.
