### 2022-01-23 #3738 @JasonGellis
Hi @Nikoleta-v3 I'm a bit unclear.
1) The version is still v1.0 so we'll keep that.
2) What am I supposed to archive on Zenodo - a copy of the repo? A pdf of the paper?
3) I understand this.
4) Will the zenodo DOI be different from the one provided by JOSS?
