### 2022-01-23 #3738 @Nikoleta-v3
1. 👍🏻
2. Please archive the repository! In case you are not familiar, Zenodo and Github work quite nicely together ([Referencing and citing content](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content)). You can create an archive by selecting the repo and a tagged release. I can see that you haven't made a tagged release of the software on GitHub: [your releases](https://github.com/alan-turing-institute/Palaeoanalytics/releases). Please make one and name it `v1.0.0` then use that one to archive the software.
3.  👍🏻
4. The Zenodo DOI will be different from the JOSS one. The Zenodo archive will appear with your paper under the `Software Archive` tab. An example paper: https://joss.theoj.org/papers/10.21105/joss.03937.

This is to ensure that an archived version of the software at the time that it was published exists 😄 Please let me know if you have any further questions, and I would be more than happy to assist you with Zenodo etc if you come across any issues!

p.s. regarding [Referencing and citing content](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content). I am sure that even if you have a Zenodo account already you can link it with your GitHub account.
