### 2022-01-24 #3738 @JasonGellis
Hi @Nikoleta-v3
1) The software has been tagged v1.0.0
2) The repo/software are now archived on Zenodo
3) The metadata on Zenodo is correct
4) https://zenodo.org/badge/latestdoi/303727518
