+++ CITATION.cff
@@ -1,12 +1,12 @@
 cff-version: 1.0.0
 message: "If you use this software, please cite it as below."
 authors:
-- family-names: "Rangel Smith"
-  given-names: "Camila"
-  orcid: "https://orcid.org/0000-0002-0227-836X"
 - family-names: "Gellis"
   given-names: "Jason"
   orcid: "https://orcid.org/0000-0002-9929-789X"
+- family-names: "Rangel Smith"
+  given-names: "Camila"
+  orcid: "https://orcid.org/0000-0002-0227-836X"
 - family-names: "Foley"
   given-names: "Robert A."
   orcid: "https://orcid.org/0000-0003-0479-3039"
