### 2022-01-24 #3738 @JasonGellis
@Nikoleta-v3 While I'm waiting to sort out hte issue with Zenodo, how do I see a copy of the paper? I followed the link above and coudn't generate a PDF.
