### 2022-01-24 #3738 @JasonGellis
@Nikoleta-v3 Ok, thank you for this. I've sorted what's happened with Zenodo. I manually changed the metadata file through their website, not realizing it was drawing infomation from the CFF file in the repo. I had hit save and thought that is was all sorted. I will update you ASAP. The paper looks good (and author names are in correct order)! Thank you for all your help.
