### 2022-01-25 #3738 @Nikoleta-v3
🤔 I am not sure which date Zenodo uses as the default - I would have thought that it would use the release date. Anyways, you can edit the upload manually if you want and you can change the date to be the same as the date of the release. Picture for example:
