### 2022-01-25 #3738 @JasonGellis
@Nikoleta-v3
1) Ok, I've changed this in the fields that you've shown. I'm guessing once I hit publish on the Zenodo page that it will change everything. Is this correct?
2) I've updated the images in the paper to all be the same size. The proof you listed above doesn't show this change. I've checked it in the Whendon Paper Preview Service, and it looks correct. These correctly sized images are now merged to the main branch of the repo.

 Is there anything else I need to do?
