### 2022-01-26 #3738 @JasonGellis
Hi @Nikoleta-v3 everything looks good on my end and I'm happy for things to move forward to publication. I have updated Zenodo and the publication date on it is now correct: https://zenodo.org/record/5898149

The DOI: 10.5281/zenodo.5898149
