### 2022-01-26 #3738 @Nikoleta-v3
👋🏻 @JasonGellis

I just saw: https://github.com/openjournals/joss-reviews/issues/3738#issuecomment-1021395240.

I can see that you fix the metadata for Zenodo 🔥  and the paper is under the Creative Commons Attribution 4.0 International License (CC BY 4.0) 😄
