### 2022-01-26 #3738 @Nikoleta-v3
🤔  I didn't quite understand the question. Do you mean where is the Zenodo information on the paper?
