### 2022-01-26 #3738 @Nikoleta-v3
👋 @openjournals/joss-eics, the paper is now ready to be accepted!

I am having an issue with the `Archive` link in the paper. It doesn't redirect you to Zenodo 🤔 I tried setting the DOI again and generating the paper but that didn't work. Could you please advice? Thank you!
