### 2022-01-28 #3738 @Nikoleta-v3
Apologies for the spam! 🙈 I just checked the `Archive` url in the final proof [openjournals/joss-papers#2911](https://github.com/openjournals/joss-papers/pull/2911) and it works just fine 👍🏻

@JasonGellis, nothing more to do on our side. Now we just wait for one of the members of the editorial board to have a final look at the  submission and accept it 🎉 Thank you again for your work.
